requirements.txt
django>=5.1.4
djangorestframework>=3.15.2
psycopg2>=2.9.10 # postgresql이랑 연결하기 위한 라이브러리
django-environ>=0.11.2
djangorestframework-simplejwt>=5.3.1

# poetry add $(cat requirements.txt)